backend/requirements.txt
# ProctoFlex AI - Backend Dependencies
# FastAPI et dépendances principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Base de données
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Validation et sérialisation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP et requêtes
httpx==0.25.2
requests==2.31.0

# Sécurité et authentification
cryptography==42.0.8
bcrypt==4.1.2

# Intelligence Artificielle et Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2

# Audio Processing (optionnel)
librosa==0.10.1
soundfile==0.12.1

# Machine Learning (optionnel - pour modèles avancés)
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Logging et monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Développement
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
